+++ src/printf_format_types.h
@@ -22,6 +22,7 @@ const std::unordered_map<std::string, Type> printf_format_types = {
   {"s", Type::string},
   {"r", Type::buffer},
   {"rx", Type::buffer},
+  {"rh", Type::buffer},
   {"c", Type::integer},
   {"d", Type::integer},
   {"u", Type::integer},